backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
inference-sdk==0.60.0
Pillow>=11.0,<12.0
python-dotenv==1.0.0

